fixed bug when issuing ?grizzly-cli` only
refactoring to also add bash completion functionality.
