+++ grizzly_cli/argparse/bashcompletion.py
@@ -1,0 +1,18 @@
+from argparse import Action
+
+__all__ = [
+    'BashCompletionAction',
+    'BashCompleteAction',
+    'BashCompletionEngine',
+]
+
+class BashCompletionAction(Action):
+    pass
+
+
+class BashCompleteAction(Action):
+    pass
+
+
+class BashCompletionEngine:
+    pass
